+++ src/main/java/com/example/myquora/repository/ThreadRepository.java
@@ -3,6 +3,6 @@ package com.example.myquora.repository;
 import com.example.myquora.model.ThreadEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ThreadRepository extends JpaRepository<ThreadEntity, Integer> {
+public interface ThreadRepository extends JpaRepository<ThreadEntity, Long> {
 
 }
